cost_accounting/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2-binary==2.9.5
djoser==2.1.0
celery==5.2.7
redis==4.3.4
